@@ -11,49 +11,49 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^16.0.0",
-    "@angular/cdk": "^16.0.0",
-    "@angular/common": "^16.0.0",
-    "@angular/compiler": "^16.0.0",
-    "@angular/core": "^16.0.0",
-    "@angular/forms": "^16.0.0",
-    "@angular/material": "^16.0.0",
-    "@angular/platform-browser": "^16.0.0",
-    "@angular/platform-browser-dynamic": "^16.0.0",
-    "@angular/router": "^16.0.0",
-    "@ngx-translate/core": "^15.0.0",
-    "@ngx-translate/http-loader": "^8.0.0",
-    "chart.js": "^4.0.0",
+    "@angular/animations": "^19.2.5",
+    "@angular/cdk": "^19.2.8",
+    "@angular/common": "^19.2.5",
+    "@angular/compiler": "^19.2.5",
+    "@angular/core": "^19.2.5",
+    "@angular/forms": "^19.2.5",
+    "@angular/material": "^19.2.8",
+    "@angular/platform-browser": "^19.2.5",
+    "@angular/platform-browser-dynamic": "^19.2.5",
+    "@angular/router": "^19.2.5",
+    "@ngx-translate/core": "^16.0.4",
+    "@ngx-translate/http-loader": "^16.0.1",
+    "chart.js": "^4.4.8",
     "crypto-js": "^4.2.0",
     "file-saver": "^2.0.5",
-    "jspdf": "^2.5.1",
+    "jspdf": "^3.0.1",
     "jwt-decode": "^4.0.0",
-    "moment": "^2.29.4",
-    "ng2-pdf-viewer": "^9.1.5",
+    "moment": "^2.30.1",
+    "ng2-pdf-viewer": "^10.4.0",
     "ngx-barcode": "^0.3.0",
     "ngx-captcha": "^13.0.0",
-    "ngx-quill": "^22.0.0",
-    "ngx-scanner-qrcode": "^1.6.6",
-    "quill": "^1.3.7",
-    "rxjs": "~7.8.0",
-    "signature_pad": "^4.1.6",
-    "tslib": "^2.3.0",
+    "ngx-quill": "^27.0.1",
+    "ngx-scanner-qrcode": "^1.7.6",
+    "quill": "^2.0.3",
+    "rxjs": "~7.8.2",
+    "signature_pad": "^5.0.7",
+    "tslib": "^2.8.1",
     "xlsx": "^0.18.5",
-    "zone.js": "~0.13.0"
+    "zone.js": "~0.15.0"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^16.0.0",
-    "@angular/cli": "~16.0.0",
-    "@angular/compiler-cli": "^16.0.0",
-    "@types/crypto-js": "^4.2.1",
+    "@angular-devkit/build-angular": "^19.2.6",
+    "@angular/cli": "^19.2.6",
+    "@angular/compiler-cli": "^19.2.5",
+    "@types/crypto-js": "^4.2.2",
     "@types/file-saver": "^2.0.7",
-    "@types/jasmine": "~4.3.0",
-    "jasmine-core": "~4.6.0",
-    "karma": "~6.4.0",
+    "@types/jasmine": "~5.1.7",
+    "jasmine-core": "~5.6.0",
+    "karma": "~6.4.4",
     "karma-chrome-launcher": "~3.2.0",
-    "karma-coverage": "~2.2.0",
+    "karma-coverage": "~2.2.1",
     "karma-jasmine": "~5.1.0",
-    "karma-jasmine-html-reporter": "~2.0.0",
-    "typescript": "~5.0.2"
+    "karma-jasmine-html-reporter": "~2.1.0",
+    "typescript": "~5.8.3"
   }
 }
