@@ -1,64 +1,25 @@
 import { NgModule } from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
-import { HttpClientModule, HttpClient, HTTP_INTERCEPTORS } from '@angular/common/http';
-import { FormsModule, ReactiveFormsModule } from '@angular/forms';
-import { RouterModule } from '@angular/router';
-
-// Translation
+import { HttpClientModule, HTTP_INTERCEPTORS } from '@angular/common/http';
+import { ServiceWorkerModule } from '@angular/service-worker';
 import { TranslateModule, TranslateLoader } from '@ngx-translate/core';
 import { TranslateHttpLoader } from '@ngx-translate/http-loader';
+import { HttpClient } from '@angular/common/http';
 
-// Material Modules
-import { MatButtonModule } from '@angular/material/button';
-import { MatCardModule } from '@angular/material/card';
-import { MatCheckboxModule } from '@angular/material/checkbox';
-import { MatDatepickerModule } from '@angular/material/datepicker';
-import { MatDialogModule } from '@angular/material/dialog';
-import { MatFormFieldModule } from '@angular/material/form-field';
-import { MatIconModule } from '@angular/material/icon';
-import { MatInputModule } from '@angular/material/input';
-import { MatListModule } from '@angular/material/list';
-import { MatMenuModule } from '@angular/material/menu';
-import { MatNativeDateModule } from '@angular/material/core';
-import { MatPaginatorModule } from '@angular/material/paginator';
-import { MatProgressBarModule } from '@angular/material/progress-bar';
-import { MatProgressSpinnerModule } from '@angular/material/progress-spinner';
-import { MatSelectModule } from '@angular/material/select';
-import { MatSidenavModule } from '@angular/material/sidenav';
-import { MatSnackBarModule } from '@angular/material/snack-bar';
-import { MatSortModule } from '@angular/material/sort';
-import { MatTableModule } from '@angular/material/table';
-import { MatTabsModule } from '@angular/material/tabs';
-import { MatToolbarModule } from '@angular/material/toolbar';
-import { MatTooltipModule } from '@angular/material/tooltip';
-
-// Rich Text Editor
-import { QuillModule } from 'ngx-quill';
-
-// PDF Viewer
-import { PdfViewerModule } from 'ng2-pdf-viewer';
-
-// App Components
-import { AppComponent } from './app.component';
 import { AppRoutingModule } from './app-routing.module';
+import { AppComponent } from './app.component';
+import { environment } from '../environments/environment';
 
-// Core Components
-import { HeaderComponent } from './core/components/header/header.component';
-import { SidenavComponent } from './core/components/sidenav/sidenav.component';
-import { FooterComponent } from './core/components/footer/footer.component';
-import { PageNotFoundComponent } from './core/components/page-not-found/page-not-found.component';
-
-// Auth Components
-import { LoginComponent } from './auth/components/login/login.component';
-import { RegisterComponent } from './auth/components/register/register.component';
-import { ForgotPasswordComponent } from './auth/components/forgot-password/forgot-password.component';
-import { ResetPasswordComponent } from './auth/components/reset-password/reset-password.component';
-
-// Interceptors
+// Core Interceptors
 import { AuthInterceptor } from './core/interceptors/auth.interceptor';
-import { TenantInterceptor } from './core/interceptors/tenant.interceptor';
 import { ErrorInterceptor } from './core/interceptors/error.interceptor';
+import { CacheInterceptor } from './core/interceptors/cache.interceptor';
+import { LoadingInterceptor } from './core/interceptors/loading.interceptor';
+import { CSRFInterceptor } from './core/interceptors/csrf.interceptor';
+
+// Shared Modules
+import { SharedModule } from './shared/shared.module';
 
 // Factory function for TranslateHttpLoader
 export function HttpLoaderFactory(http: HttpClient) {
@@ -67,69 +28,39 @@ export function HttpLoaderFactory(http: HttpClient) {
 
 @NgModule({
   declarations: [
-    AppComponent,
-    HeaderComponent,
-    SidenavComponent,
-    FooterComponent,
-    PageNotFoundComponent,
-    LoginComponent,
-    RegisterComponent,
-    ForgotPasswordComponent,
-    ResetPasswordComponent
+    AppComponent
   ],
   imports: [
-    BrowserModule,
+    BrowserModule.withServerTransition({ appId: 'adalalegalis-app' }), // Enable server-side rendering
     BrowserAnimationsModule,
     HttpClientModule,
-    FormsModule,
-    ReactiveFormsModule,
-    RouterModule,
     AppRoutingModule,
+    SharedModule,
     
-    // Translation
+    // Service Worker for PWA support
+    ServiceWorkerModule.register('ngsw-worker.js', {
+      enabled: environment.production,
+      // Register the ServiceWorker as soon as the application is stable
+      registrationStrategy: 'registerWhenStable:30000'
+    }),
+    
+    // Translation module
     TranslateModule.forRoot({
+      defaultLanguage: 'ar',
       loader: {
         provide: TranslateLoader,
         useFactory: HttpLoaderFactory,
         deps: [HttpClient]
-      },
-      defaultLanguage: 'ar'
-    }),
-    
-    // Material Modules
-    MatButtonModule,
-    MatCardModule,
-    MatCheckboxModule,
-    MatDatepickerModule,
-    MatDialogModule,
-    MatFormFieldModule,
-    MatIconModule,
-    MatInputModule,
-    MatListModule,
-    MatMenuModule,
-    MatNativeDateModule,
-    MatPaginatorModule,
-    MatProgressBarModule,
-    MatProgressSpinnerModule,
-    MatSelectModule,
-    MatSidenavModule,
-    MatSnackBarModule,
-    MatSortModule,
-    MatTableModule,
-    MatTabsModule,
-    MatToolbarModule,
-    MatTooltipModule,
-    
-    // Rich Text Editor
-    QuillModule.forRoot(),
-    
-    // PDF Viewer
-    PdfViewerModule
+      }
+    })
   ],
   providers: [
+    // HTTP Interceptors
     { provide: HTTP_INTERCEPTORS, useClass: AuthInterceptor, multi: true },
-    { provide: HTTP_INTERCEPTORS, useClass: TenantInterceptor, multi: true },
-    { provide: HTTP_INTERCEPTORS, useClass: ErrorInterceptor, multi: true }
+    { provide: HTTP_INTERCEPTORS, useClass: ErrorInterceptor, multi: true },
+    { provide: HTTP_INTERCEPTORS, useClass: CacheInterceptor, multi: true },
+    { provide: HTTP_INTERCEPTORS, useClass: LoadingInterceptor, multi: true },
+    { provide: HTTP_INTERCEPTORS, useClass: CSRFInterceptor, multi: true }
   ],
   bootstrap: [AppComponent]
 })
