@@ -1,308 +1,376 @@
 /**
- * @file error.middleware.js
- * @description وسيط معالجة الأخطاء المركزي للتطبيق
- * يوفر هذا الملف آلية موحدة لمعالجة الأخطاء وإدارتها في جميع أنحاء التطبيق
+ * وحدة معالجة الأخطاء المركزية
+ * 
+ * هذا الملف يوفر آليات موحدة لمعالجة الأخطاء في التطبيق، بما في ذلك:
+ * - معالجة أخطاء HTTP
+ * - معالجة أخطاء التحقق من الصحة
+ * - معالجة أخطاء قاعدة البيانات
+ * - تسجيل الأخطاء وتتبعها
+ * - إرسال إشعارات الأخطاء
  */
 
-const { ValidationError } = require('express-validation');
-const { isCelebrateError } = require('celebrate');
-const winston = require('winston');
-const Sentry = require('@sentry/node');
-const { environment } = require('../config/environment');
-
-// تكوين Sentry لتتبع الأخطاء إذا كان مكوناً
-if (environment.sentry && environment.sentry.dsn) {
-  Sentry.init({
-    dsn: environment.sentry.dsn,
-    environment: process.env.NODE_ENV,
-    integrations: [
-      new Sentry.Integrations.Http({ tracing: true }),
-      new Sentry.Integrations.Express({ app: null }),
-    ],
-    tracesSampleRate: 1.0,
-  });
-}
-
-// إنشاء مسجل Winston
-const logger = winston.createLogger({
-  level: process.env.LOG_LEVEL || 'info',
-  format: winston.format.combine(
-    winston.format.timestamp(),
-    winston.format.json()
-  ),
-  defaultMeta: { service: 'adalalegalis-api' },
-  transports: [
-    new winston.transports.Console({
-      format: winston.format.combine(
-        winston.format.colorize(),
-        winston.format.simple()
-      ),
-    }),
-  ],
-});
-
-// إضافة نقل ملف إذا كان تمكين ملف السجل صحيحًا
-if (process.env.LOG_FILE_ENABLED === 'true') {
-  logger.add(
-    new winston.transports.File({
-      filename: process.env.LOG_FILE_PATH || 'logs/app.log',
-      maxsize: process.env.LOG_MAX_SIZE || '10m',
-      maxFiles: process.env.LOG_MAX_FILES || 7,
-    })
-  );
-}
+const { logger } = require('../utils/logger');
+const { ValidationError } = require('sequelize');
+const config = require('../config/app.config');
 
 /**
- * أنواع الأخطاء المعرفة في التطبيق
+ * تصنيفات الأخطاء المختلفة
  */
 const ErrorTypes = {
-  VALIDATION: 'ValidationError',
-  AUTHENTICATION: 'AuthenticationError',
-  AUTHORIZATION: 'AuthorizationError',
-  NOT_FOUND: 'NotFoundError',
-  CONFLICT: 'ConflictError',
-  INTERNAL: 'InternalError',
-  EXTERNAL_SERVICE: 'ExternalServiceError',
-  DATABASE: 'DatabaseError',
-  BUSINESS_LOGIC: 'BusinessLogicError',
+  VALIDATION: 'VALIDATION_ERROR',
+  NOT_FOUND: 'NOT_FOUND_ERROR',
+  UNAUTHORIZED: 'UNAUTHORIZED_ERROR',
+  FORBIDDEN: 'FORBIDDEN_ERROR',
+  DATABASE: 'DATABASE_ERROR',
+  EXTERNAL_SERVICE: 'EXTERNAL_SERVICE_ERROR',
+  RATE_LIMIT: 'RATE_LIMIT_ERROR',
+  INTERNAL: 'INTERNAL_SERVER_ERROR',
+  BAD_REQUEST: 'BAD_REQUEST_ERROR',
+  CONFLICT: 'CONFLICT_ERROR'
 };
 
 /**
- * فئة الخطأ الأساسية للتطبيق
+ * رموز حالة HTTP المقابلة لأنواع الأخطاء
+ */
+const ErrorStatusCodes = {
+  [ErrorTypes.VALIDATION]: 400,
+  [ErrorTypes.NOT_FOUND]: 404,
+  [ErrorTypes.UNAUTHORIZED]: 401,
+  [ErrorTypes.FORBIDDEN]: 403,
+  [ErrorTypes.DATABASE]: 500,
+  [ErrorTypes.EXTERNAL_SERVICE]: 503,
+  [ErrorTypes.RATE_LIMIT]: 429,
+  [ErrorTypes.INTERNAL]: 500,
+  [ErrorTypes.BAD_REQUEST]: 400,
+  [ErrorTypes.CONFLICT]: 409
+};
+
+/**
+ * فئة الخطأ المخصصة للتطبيق
  */
 class AppError extends Error {
-  constructor(message, type, statusCode, details = null, isOperational = true) {
+  constructor(message, type = ErrorTypes.INTERNAL, details = null, originalError = null) {
     super(message);
-    this.name = type;
-    this.statusCode = statusCode;
+    this.name = 'AppError';
+    this.type = type;
     this.details = details;
-    this.isOperational = isOperational;
+    this.originalError = originalError;
     this.timestamp = new Date().toISOString();
+    this.statusCode = ErrorStatusCodes[type] || 500;
     
-    Error.captureStackTrace(this, this.constructor);
+    // تسجيل معلومات إضافية للتصحيح
+    if (originalError) {
+      this.stack = originalError.stack;
+    } else {
+      Error.captureStackTrace(this, this.constructor);
+    }
+  }
+  
+  /**
+   * تحويل الخطأ إلى كائن JSON للاستجابة
+   */
+  toJSON() {
+    const response = {
+      status: 'error',
+      type: this.type,
+      message: this.message,
+      timestamp: this.timestamp
+    };
+    
+    // إضافة التفاصيل إذا كانت موجودة
+    if (this.details) {
+      response.details = this.details;
+    }
+    
+    // إضافة معلومات التصحيح في بيئة التطوير
+    if (config.nodeEnv === 'development' && this.originalError) {
+      response.debug = {
+        originalMessage: this.originalError.message,
+        stack: this.originalError.stack
+      };
+    }
+    
+    return response;
+  }
+  
+  /**
+   * إنشاء خطأ تحقق من الصحة
+   */
+  static validation(message, details) {
+    return new AppError(
+      message || 'بيانات غير صالحة',
+      ErrorTypes.VALIDATION,
+      details
+    );
+  }
+  
+  /**
+   * إنشاء خطأ غير موجود
+   */
+  static notFound(message, details) {
+    return new AppError(
+      message || 'المورد غير موجود',
+      ErrorTypes.NOT_FOUND,
+      details
+    );
+  }
+  
+  /**
+   * إنشاء خطأ غير مصرح به
+   */
+  static unauthorized(message, details) {
+    return new AppError(
+      message || 'غير مصرح به',
+      ErrorTypes.UNAUTHORIZED,
+      details
+    );
+  }
+  
+  /**
+   * إنشاء خطأ ممنوع
+   */
+  static forbidden(message, details) {
+    return new AppError(
+      message || 'الوصول ممنوع',
+      ErrorTypes.FORBIDDEN,
+      details
+    );
+  }
+  
+  /**
+   * إنشاء خطأ قاعدة بيانات
+   */
+  static database(message, originalError) {
+    return new AppError(
+      message || 'حدث خطأ في قاعدة البيانات',
+      ErrorTypes.DATABASE,
+      null,
+      originalError
+    );
+  }
+  
+  /**
+   * إنشاء خطأ خدمة خارجية
+   */
+  static externalService(message, details, originalError) {
+    return new AppError(
+      message || 'فشل الاتصال بالخدمة الخارجية',
+      ErrorTypes.EXTERNAL_SERVICE,
+      details,
+      originalError
+    );
+  }
+  
+  /**
+   * إنشاء خطأ تجاوز الحد
+   */
+  static rateLimit(message, details) {
+    return new AppError(
+      message || 'تم تجاوز حد معدل الطلبات',
+      ErrorTypes.RATE_LIMIT,
+      details
+    );
+  }
+  
+  /**
+   * إنشاء خطأ طلب غير صالح
+   */
+  static badRequest(message, details) {
+    return new AppError(
+      message || 'طلب غير صالح',
+      ErrorTypes.BAD_REQUEST,
+      details
+    );
+  }
+  
+  /**
+   * إنشاء خطأ تعارض
+   */
+  static conflict(message, details) {
+    return new AppError(
+      message || 'حدث تعارض مع الحالة الحالية للمورد',
+      ErrorTypes.CONFLICT,
+      details
+    );
   }
 }
 
 /**
- * إنشاء أخطاء محددة للتطبيق
+ * وسيط معالجة الأخطاء المركزي
+ * يتعامل مع جميع الأخطاء ويعيد استجابة منسقة
  */
-const createAppErrors = () => {
-  return {
-    ValidationError: (message, details) => 
-      new AppError(message, ErrorTypes.VALIDATION, 400, details),
-    
-    AuthenticationError: (message, details) => 
-      new AppError(message, ErrorTypes.AUTHENTICATION, 401, details),
-    
-    AuthorizationError: (message, details) => 
-      new AppError(message, ErrorTypes.AUTHORIZATION, 403, details),
-    
-    NotFoundError: (message, details) => 
-      new AppError(message, ErrorTypes.NOT_FOUND, 404, details),
-    
-    ConflictError: (message, details) => 
-      new AppError(message, ErrorTypes.CONFLICT, 409, details),
-    
-    InternalError: (message, details, isOperational = true) => 
-      new AppError(message, ErrorTypes.INTERNAL, 500, details, isOperational),
-    
-    ExternalServiceError: (message, details) => 
-      new AppError(message, ErrorTypes.EXTERNAL_SERVICE, 502, details),
-    
-    DatabaseError: (message, details) => 
-      new AppError(message, ErrorTypes.DATABASE, 500, details),
-    
-    BusinessLogicError: (message, details) => 
-      new AppError(message, ErrorTypes.BUSINESS_LOGIC, 422, details),
-  };
-};
-
-// إنشاء أخطاء التطبيق
-const AppErrors = createAppErrors();
-
-/**
- * تحويل أخطاء المكتبات الخارجية إلى أخطاء التطبيق
- * @param {Error} error الخطأ الأصلي
- * @returns {AppError} خطأ التطبيق
- */
-const convertError = (error) => {
-  // تحويل أخطاء التحقق من الصحة
-  if (error instanceof ValidationError || isCelebrateError(error)) {
-    return AppErrors.ValidationError(
-      'خطأ في التحقق من صحة البيانات',
-      error.details || error.message
-    );
-  }
+const errorHandler = (err, req, res, next) => {
+  let error = err;
   
-  // تحويل أخطاء Mongoose
-  if (error.name === 'MongooseError' || error.name === 'MongoError') {
-    return AppErrors.DatabaseError(
-      'خطأ في قاعدة البيانات',
-      error.message
-    );
-  }
-  
-  // تحويل أخطاء JsonWebToken
-  if (error.name === 'JsonWebTokenError' || error.name === 'TokenExpiredError') {
-    return AppErrors.AuthenticationError(
-      'خطأ في المصادقة',
-      error.message
-    );
-  }
-  
-  // إذا كان الخطأ غير معروف، إرجاع خطأ داخلي
-  return AppErrors.InternalError(
-    'حدث خطأ داخلي',
-    error.message,
-    false
-  );
-};
-
-/**
- * تسجيل الخطأ في السجلات وخدمات التتبع
- * @param {Error} error الخطأ المراد تسجيله
- * @param {Request} req طلب Express
- */
-const logError = (error, req) => {
-  // تسجيل الخطأ في Winston
+  // تسجيل الخطأ
   logger.error({
-    message: error.message,
-    type: error.name,
-    statusCode: error.statusCode,
-    details: error.details,
-    stack: error.stack,
-    path: req.path,
+    message: err.message,
+    stack: err.stack,
+    requestId: req.id,
+    url: req.originalUrl,
     method: req.method,
     ip: req.ip,
-    userId: req.user ? req.user.id : null,
+    userId: req.user ? req.user.id : 'anonymous'
   });
   
-  // تسجيل الخطأ في Sentry إذا كان مكوناً وكان الخطأ غير تشغيلي
-  if (environment.sentry && environment.sentry.dsn && !error.isOperational) {
-    Sentry.withScope((scope) => {
-      scope.setUser({ id: req.user ? req.user.id : 'anonymous' });
-      scope.setTag('path', req.path);
-      scope.setTag('method', req.method);
-      scope.setExtra('details', error.details);
-      Sentry.captureException(error);
+  // تحويل أخطاء Sequelize إلى أخطاء تطبيق
+  if (err instanceof ValidationError) {
+    const validationErrors = {};
+    
+    err.errors.forEach(error => {
+      validationErrors[error.path] = error.message;
     });
-  }
-};
-
-/**
- * إرسال استجابة الخطأ إلى العميل
- * @param {Error} error الخطأ المراد إرساله
- * @param {Request} req طلب Express
- * @param {Response} res استجابة Express
- */
-const sendErrorResponse = (error, req, res) => {
-  // تحديد ما إذا كان يجب إرسال تفاصيل الخطأ
-  const shouldSendDetails = process.env.NODE_ENV !== 'production' || error.isOperational;
-  
-  // إنشاء استجابة الخطأ
-  const errorResponse = {
-    status: 'error',
-    message: error.message,
-    code: error.name,
-    requestId: req.id,
-  };
-  
-  // إضافة تفاصيل الخطأ إذا كان مسموحًا
-  if (shouldSendDetails && error.details) {
-    errorResponse.details = error.details;
+    
+    error = AppError.validation('خطأ في التحقق من صحة البيانات', validationErrors);
   }
   
-  // إضافة تتبع المكدس في بيئة التطوير
-  if (process.env.NODE_ENV === 'development') {
-    errorResponse.stack = error.stack;
+  // تحويل أخطاء JWT إلى أخطاء غير مصرح بها
+  if (err.name === 'JsonWebTokenError' || err.name === 'TokenExpiredError') {
+    error = AppError.unauthorized('رمز المصادقة غير صالح أو منتهي الصلاحية');
+  }
+  
+  // تحويل أخطاء Mongoose إلى أخطاء تطبيق
+  if (err.name === 'CastError') {
+    error = AppError.badRequest(`قيمة غير صالحة: ${err.value}`);
+  }
+  
+  // تحويل أخطاء Mongoose الفريدة إلى أخطاء تعارض
+  if (err.code === 11000) {
+    const field = Object.keys(err.keyValue)[0];
+    error = AppError.conflict(`القيمة ${err.keyValue[field]} مستخدمة بالفعل للحقل ${field}`);
+  }
+  
+  // إذا لم يكن الخطأ من نوع AppError، قم بتحويله
+  if (!(error instanceof AppError)) {
+    error = new AppError(
+      config.nodeEnv === 'production' ? 'حدث خطأ في الخادم' : error.message,
+      ErrorTypes.INTERNAL,
+      null,
+      error
+    );
   }
   
   // إرسال الاستجابة
-  res.status(error.statusCode || 500).json(errorResponse);
+  res.status(error.statusCode).json(error.toJSON());
+  
+  // إرسال إشعار للمسؤولين في حالة الأخطاء الخطيرة
+  if (error.statusCode >= 500) {
+    notifyAdminsAboutError(error, req);
+  }
 };
 
 /**
- * وسيط معالجة الأخطاء
- * @param {Error} err الخطأ المراد معالجته
- * @param {Request} req طلب Express
- * @param {Response} res استجابة Express
- * @param {Function} next دالة التالي
+ * وسيط معالجة المسارات غير الموجودة
  */
-const errorMiddleware = (err, req, res, next) => {
-  // تحويل الخطأ إلى خطأ التطبيق إذا لم يكن كذلك بالفعل
-  const error = err instanceof AppError ? err : convertError(err);
-  
-  // تسجيل الخطأ
-  logError(error, req);
-  
-  // إرسال استجابة الخطأ
-  sendErrorResponse(error, req, res);
+const notFoundHandler = (req, res, next) => {
+  next(AppError.notFound(`المسار غير موجود: ${req.originalUrl}`));
 };
 
 /**
- * وسيط معالجة الطرق غير الموجودة
- * @param {Request} req طلب Express
- * @param {Response} res استجابة Express
- * @param {Function} next دالة التالي
+ * وسيط معالجة أخطاء التحقق من الصحة
  */
-const notFoundMiddleware = (req, res, next) => {
-  const error = AppErrors.NotFoundError(
-    `لم يتم العثور على المسار: ${req.originalUrl}`,
-    { path: req.originalUrl }
-  );
+const validationErrorHandler = (err, req, res, next) => {
+  if (err && err.error && err.error.isJoi) {
+    const validationErrors = {};
+    
+    err.error.details.forEach(detail => {
+      validationErrors[detail.context.key] = detail.message;
+    });
+    
+    return next(AppError.validation('خطأ في التحقق من صحة البيانات', validationErrors));
+  }
   
-  next(error);
+  next(err);
 };
 
 /**
- * وسيط معالجة الأخطاء غير المتوقعة
- * @param {Error} error الخطأ غير المتوقع
+ * إشعار المسؤولين بالأخطاء الخطيرة
+ * يمكن تنفيذ هذه الوظيفة لإرسال بريد إلكتروني أو إشعار Slack أو غيرها
  */
-const handleUncaughtException = (error) => {
+const notifyAdminsAboutError = (error, req) => {
+  // في بيئة الإنتاج، يمكن تنفيذ إرسال إشعارات عبر البريد الإلكتروني أو Slack
+  if (config.nodeEnv === 'production') {
+    // مثال: إرسال بريد إلكتروني للمسؤولين
+    // emailService.sendErrorNotification({
+    //   subject: `خطأ خطير في التطبيق: ${error.type}`,
+    //   error: error,
+    //   request: {
+    //     url: req.originalUrl,
+    //     method: req.method,
+    //     ip: req.ip,
+    //     userId: req.user ? req.user.id : 'anonymous',
+    //     timestamp: new Date().toISOString()
+    //   }
+    // });
+    
+    logger.warn('تم تسجيل خطأ خطير، يجب تنفيذ آلية إشعار المسؤولين');
+  }
+};
+
+/**
+ * وسيط تسجيل الطلبات
+ * يسجل معلومات عن كل طلب قبل معالجته
+ */
+const requestLogger = (req, res, next) => {
+  const start = Date.now();
+  
+  // تسجيل بداية الطلب
+  logger.info({
+    message: 'طلب وارد',
+    method: req.method,
+    url: req.originalUrl,
+    ip: req.ip,
+    requestId: req.id,
+    userId: req.user ? req.user.id : 'anonymous'
+  });
+  
+  // تسجيل نهاية الطلب ووقت الاستجابة
+  res.on('finish', () => {
+    const duration = Date.now() - start;
+    const level = res.statusCode >= 400 ? 'warn' : 'info';
+    
+    logger[level]({
+      message: 'انتهاء الطلب',
+      method: req.method,
+      url: req.originalUrl,
+      status: res.statusCode,
+      duration: `${duration}ms`,
+      requestId: req.id,
+      userId: req.user ? req.user.id : 'anonymous'
+    });
+  });
+  
+  next();
+};
+
+/**
+ * وسيط التعامل مع الأخطاء غير المتوقعة
+ * يتعامل مع الأخطاء غير المتوقعة مثل الاستثناءات غير المعالجة
+ */
+const unexpectedErrorHandler = (error) => {
   logger.error({
     message: 'خطأ غير متوقع',
     error: error.message,
-    stack: error.stack,
+    stack: error.stack
   });
   
-  // تسجيل الخطأ في Sentry إذا كان مكوناً
-  if (environment.sentry && environment.sentry.dsn) {
-    Sentry.captureException(error);
+  // في بيئة الإنتاج، يمكن إعادة تشغيل التطبيق أو اتخاذ إجراء آخر
+  if (config.nodeEnv === 'production') {
+    // يمكن تنفيذ آلية إعادة تشغيل هنا
+    logger.warn('حدث خطأ غير متوقع في بيئة الإنتاج، يجب تنفيذ آلية التعافي');
   }
-  
-  // إنهاء العملية بأمان بعد تسجيل الخطأ
-  process.exit(1);
 };
 
-/**
- * وسيط معالجة الوعود المرفوضة غير المعالجة
- * @param {Error} error الخطأ غير المعالج
- */
-const handleUnhandledRejection = (error) => {
-  logger.error({
-    message: 'وعد مرفوض غير معالج',
-    error: error.message,
-    stack: error.stack,
-  });
-  
-  // تسجيل الخطأ في Sentry إذا كان مكوناً
-  if (environment.sentry && environment.sentry.dsn) {
-    Sentry.captureException(error);
-  }
-  
-  // لا ننهي العملية هنا لأن الوعود المرفوضة غير المعالجة لا تترك التطبيق في حالة غير متسقة
-};
+// تسجيل معالجي الأخطاء غير المتوقعة على مستوى العملية
+process.on('uncaughtException', unexpectedErrorHandler);
+process.on('unhandledRejection', unexpectedErrorHandler);
 
-// تسجيل معالجات الأخطاء العالمية
-process.on('uncaughtException', handleUncaughtException);
-process.on('unhandledRejection', handleUnhandledRejection);
-
-// تصدير الوسطاء والأخطاء
 module.exports = {
-  errorMiddleware,
-  notFoundMiddleware,
-  AppErrors,
-  ErrorTypes,
   AppError,
-  logger,
+  ErrorTypes,
+  errorHandler,
+  notFoundHandler,
+  validationErrorHandler,
+  requestLogger,
+  unexpectedErrorHandler
 };
